changed from double to decimal for duration,distance,lats,longs added annotation option to route request

+++ Src/Osrm.Client/Models/RouteStep.cs
@@ -12,10 +12,10 @@ namespace Osrm.Client.Models
     public class RouteStep
     {
         [JsonPropertyName("distance")]
-        public double Distance { get; set; }
+        public decimal Distance { get; set; }
 
         [JsonPropertyName("duration")]
-        public double Duration { get; set; }
+        public decimal Duration { get; set; }
 
         [JsonPropertyName("geometry")]
         public string GeometryStr { get; set; }
@@ -29,7 +29,7 @@ namespace Osrm.Client.Models
                     return new Location[0];
                 }
 
-                return OsrmPolylineConverter.Decode(GeometryStr, 1E5)
+                return OsrmPolylineConverter.Decode(GeometryStr, 1E5m)
                     .ToArray();
             }
         }
